requirements.txt
Flask==0.10.1
Flask-MySQL==1.3
Flask-MySQLdb==0.2.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
mysqlclient==1.3.7
vboxapi==1.0
virtualenv==13.1.0
Werkzeug==0.11.3
wheel==0.24.0
